Add FocusHighlight wiget, moves Focusable into widgets from material
