+++ packages/flutter/lib/material.dart
@@ -58,7 +58,7 @@ export 'src/material/floating_action_button.dart';
 export 'src/material/floating_action_button_location.dart';
 export 'src/material/floating_action_button_theme.dart';
 export 'src/material/flutter_logo.dart';
-export 'src/material/focusable.dart';
+export 'src/material/focus_highlight.dart';
 export 'src/material/grid_tile.dart';
 export 'src/material/grid_tile_bar.dart';
 export 'src/material/icon_button.dart';
