+++ packages/flutter/lib/src/material/theme_data.dart
@@ -18,7 +18,7 @@ import 'color_scheme.dart';
 import 'colors.dart';
 import 'dialog_theme.dart';
 import 'floating_action_button_theme.dart';
-import 'focusable.dart';
+import 'focus_highlight.dart';
 import 'ink_splash.dart';
 import 'ink_well.dart' show InteractiveInkFeatureFactory;
 import 'input_decorator.dart';
@@ -164,7 +164,7 @@ class ThemeData extends Diagnosticable {
     FloatingActionButtonThemeData floatingActionButtonTheme,
     Typography typography,
     CupertinoThemeData cupertinoOverrideTheme,
-    FocusableThemeData focusableTheme,
+    FocusHighlightThemeData focusableTheme,
   }) {
     brightness ??= Brightness.light;
     final bool isDark = brightness == Brightness.dark;
@@ -264,7 +264,7 @@ class ThemeData extends Diagnosticable {
     floatingActionButtonTheme ??= const FloatingActionButtonThemeData();
     cupertinoOverrideTheme = cupertinoOverrideTheme?.noDefault();
 
-    focusableTheme ??= const FocusableThemeData();
+    focusableTheme ??= const FocusHighlightThemeData();
 
     return ThemeData.raw(
       brightness: brightness,
@@ -318,7 +318,7 @@ class ThemeData extends Diagnosticable {
       floatingActionButtonTheme: floatingActionButtonTheme,
       typography: typography,
       cupertinoOverrideTheme: cupertinoOverrideTheme,
-      focusableTheme: focusableTheme,
+      focusHighlightTheme: focusableTheme,
     );
   }
 
@@ -384,7 +384,7 @@ class ThemeData extends Diagnosticable {
     @required this.floatingActionButtonTheme,
     @required this.typography,
     @required this.cupertinoOverrideTheme,
-    @required this.focusableTheme,
+    @required this.focusHighlightTheme,
   }) : assert(brightness != null),
        assert(primaryColor != null),
        assert(primaryColorBrightness != null),
@@ -434,7 +434,7 @@ class ThemeData extends Diagnosticable {
        assert(dialogTheme != null),
        assert(floatingActionButtonTheme != null),
        assert(typography != null),
-       assert(focusableTheme != null);
+       assert(focusHighlightTheme != null);
 
   // Warning: make sure these properties are in the exact same order as in
   // hashValues() and in the raw constructor and in the order of fields in
@@ -697,11 +697,11 @@ class ThemeData extends Diagnosticable {
 
   /// Defines the appearance of [Focusable]s.
   ///
-  /// This property is used to define the value of [ThemeData.focusableTheme].
+  /// This property is used to define the value of [ThemeData.focusHighlightTheme].
   ///
   /// The [Focusable] widget uses the current focusable theme to initialize some
   /// [Focusable] properties (like [Focusable.focusedDecoration]).
-  final FocusableThemeData focusableTheme;
+  final FocusHighlightThemeData focusHighlightTheme;
 
   /// Creates a copy of this theme but with the given fields replaced with the new values.
   ThemeData copyWith({
@@ -756,7 +756,7 @@ class ThemeData extends Diagnosticable {
     FloatingActionButtonThemeData floatingActionButtonTheme,
     Typography typography,
     CupertinoThemeData cupertinoOverrideTheme,
-    FocusableThemeData focusableTheme,
+    FocusHighlightThemeData focusableTheme,
   }) {
     cupertinoOverrideTheme = cupertinoOverrideTheme?.noDefault();
     return ThemeData.raw(
@@ -811,7 +811,7 @@ class ThemeData extends Diagnosticable {
       floatingActionButtonTheme: floatingActionButtonTheme ?? this.floatingActionButtonTheme,
       typography: typography ?? this.typography,
       cupertinoOverrideTheme: cupertinoOverrideTheme ?? this.cupertinoOverrideTheme,
-      focusableTheme: focusableTheme ?? this.focusableTheme,
+      focusHighlightTheme: focusableTheme ?? this.focusHighlightTheme,
     );
   }
 
@@ -944,7 +944,7 @@ class ThemeData extends Diagnosticable {
       floatingActionButtonTheme: FloatingActionButtonThemeData.lerp(a.floatingActionButtonTheme, b.floatingActionButtonTheme, t),
       typography: Typography.lerp(a.typography, b.typography, t),
       cupertinoOverrideTheme: t < 0.5 ? a.cupertinoOverrideTheme : b.cupertinoOverrideTheme,
-      focusableTheme: FocusableThemeData.lerp(a.focusableTheme, b.focusableTheme, t),
+      focusHighlightTheme: FocusHighlightThemeData.lerp(a.focusHighlightTheme, b.focusHighlightTheme, t),
     );
   }
