+++ packages/flutter/lib/widgets.dart
@@ -35,6 +35,7 @@ export 'src/widgets/fade_in_image.dart';
 export 'src/widgets/focus_manager.dart';
 export 'src/widgets/focus_scope.dart';
 export 'src/widgets/focus_traversal.dart';
+export 'src/widgets/focusable.dart';
 export 'src/widgets/form.dart';
 export 'src/widgets/framework.dart';
 export 'src/widgets/gesture_detector.dart';
